+++ packages/eslint-plugin-stories/src/index.ts
@@ -1,11 +1,13 @@
 import noExtResourcesInStories from './rules/no-ext-resources-in-stories';
 import noJestInStories from './rules/no-jest-in-stories';
+import noStoryDecorators from './rules/no-story-decorators';
 import noTopLevelStoryArgs from './rules/no-top-level-story-args';
 import storiesDefaultExport from './rules/stories-default-export';
 
 const rules = {
   'no-ext-resources-in-stories': noExtResourcesInStories,
   'no-jest-in-stories': noJestInStories,
+  'no-story-decorators': noStoryDecorators,
   'no-top-level-story-args': noTopLevelStoryArgs,
   'stories-default-export': storiesDefaultExport,
 };
@@ -13,6 +15,7 @@ const rules = {
 const recommendedRules = {
   '@chanzuckerberg/stories/no-ext-resources-in-stories': 'error',
   '@chanzuckerberg/stories/no-jest-in-stories': 'error',
+  '@chanzuckerberg/stories/no-story-decorators': 'error',
   '@chanzuckerberg/stories/no-top-level-story-args': 'error',
   '@chanzuckerberg/stories/stories-default-export': 'error',
 };